requirements.txt
# ====================
# CORE DEPENDENCIES
# ====================
Django==4.2.7
python-dotenv==1.0.0
dj-database-url==2.0.0

# ====================
# DATABASE
# ====================
psycopg2-binary==2.9.9

# ====================
# CONTENT & MEDIA MANAGEMENT
# ====================
# Rich Text Editors
django-ckeditor==6.7.0
django-summernote==0.8.20.0

# Content Enhancement
django-taggit==4.0.0
Pillow==9.5.0

# Cloud Media Storage & Optimization
cloudinary==1.38.0
django-cloudinary-storage==0.3.0

# ====================
# USER MANAGEMENT
# ====================
django-allauth==0.58.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10

# ====================
# PERFORMANCE & SEO
# ====================
django-compressor==4.1
# django-seo-js==2.1.0  # Uncomment if needed
# django-sitemaps==3.0.0  # Uncomment if needed

# ====================
# PRODUCTION & DEPLOYMENT
# ====================
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.0  # Keep if you need S3 for other files
boto3==1.28.57

# ====================
# INTERNATIONALIZATION
# ====================
django-countries==7.5